+++ CHANGELOG.md
@@ -55,6 +55,7 @@ _This changelog only relates to the `raffia` crate._
 - Changed AST of nesting selector.
 - Reject duplicated Sass flag in variable declaration.
 - Added support of Less property variable.
+- Added support of Less escaped string.
 
 ## v0.2.0